Save several queries by simplifying the 'misc' table format.
This also places the Attention Box title in the database for consistency and removes the leftovers of being able to have the entire $config in the database (which was horrendously broken anyway).
Note that in case the option to have $config in the database reappears properly, it should not use this new format; and it should go into a different table, like 'boardconfig'.